同步smp bsp结构 去除rt_hw_mb定义
