+++ bsp/qemu-vexpress-a9/drivers/board.c
@@ -19,10 +19,8 @@
 
 #define SYS_CTRL                        __REG32(REALVIEW_SCTL_BASE)
 
-void idle_wfi(void)
-{
-    asm volatile ("wfi");
-}
+extern void idle_wfi(void);
+extern void rt_hw_ipi_handler_install(int ipi_vector, rt_isr_handler_t ipi_isr_handler);
 
 /**
  * This function will initialize beaglebone board
@@ -40,52 +38,8 @@ void rt_hw_board_init(void)
     rt_thread_idle_sethook(idle_wfi);
 
 #ifdef RT_USING_SMP
-    /* install IPI interrupt */
-    rt_hw_interrupt_install(RT_SCHEDULE_IPI_IRQ, rt_scheduler_ipi_handler, RT_NULL, "ipi");
-    rt_hw_interrupt_umask(RT_SCHEDULE_IPI_IRQ);
+    /* install IPI handle */
+    rt_hw_ipi_handler_install(RT_SCHEDULE_IPI, rt_scheduler_ipi_handler);
 #endif
 }
 
-#ifdef RT_USING_SMP
-void rt_hw_spin_lock(rt_hw_spinlock_t *lock)
-{
-    unsigned long tmp;
-    unsigned long newval;
-    rt_hw_spinlock_t lockval;
-
-    __asm__ __volatile__(
-            "pld [%0]"
-            ::"r"(&lock->slock)
-            );
-
-    __asm__ __volatile__(
-            "1: ldrex   %0, [%3]\n"
-            "   add %1, %0, %4\n"
-            "   strex   %2, %1, [%3]\n"
-            "   teq %2, #0\n"
-            "   bne 1b"
-            : "=&r" (lockval), "=&r" (newval), "=&r" (tmp)
-            : "r" (&lock->slock), "I" (1 << 16)
-            : "cc");
-
-    while (lockval.tickets.next != lockval.tickets.owner) {
-        __asm__ __volatile__("wfe":::"memory");
-        lockval.tickets.owner = *(volatile unsigned short *)(&lock->tickets.owner);
-    }
-
-    __asm__ volatile ("dmb":::"memory");
-}
-
-void rt_hw_spin_unlock(rt_hw_spinlock_t *lock)
-{
-    __asm__ volatile ("dmb":::"memory");
-    lock->tickets.owner++;
-    __asm__ volatile ("dsb ishst\nsev":::"memory");
-}
-
-void rt_hw_mb(void)
-{
-    __asm__ volatile ("dmb":::"memory");
-}
-
-#endif /*RT_USING_SMP*/
